% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbard.R
\name{geom_errorbard}
\alias{geom_errorbard}
\title{Diagonal error bar}
\usage{
geom_errorbard(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{xcenter}{Aesthetic mapping for the center x coordinate of error bar. Use this instead of \code{x}.}

\item{ycenter}{Aesthetic mapping for the center y coordinate of error bar. Use this instead of \code{y}.}

\item{latitude}{Aesthetic mapping for the total breadth of the error interval. Use this instead of \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}.}

\item{angle}{Angle of error bars in degrees}
}
\description{
Draw a diagonal (or arbitrarily oriented) error bar in a \code{ggplot2}, inherits from \link[ggplot2:geom_segment]{ggplot2::geom_segment}.
}
\seealso{
[\link[ggplot2:geom_segment]{ggplot2::geom_segment}]
}
