% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbard.R
\docType{data}
\name{GeomErrorbarD}
\alias{GeomErrorbarD}
\title{GeomErrorBarD}
\format{
An object of class \code{GeomErrorbarD} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomErrorbarD
}
\description{
\code{GeomErrorBarD} proto type definition based on \code{ggplot2::GeomSegment}.
}
\keyword{datasets}
