% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbard.R
\name{rotate_around}
\alias{rotate_around}
\title{Rotate a point around another point in 2D.}
\usage{
rotate_around(x, y, cx, cy, angle_deg)
}
\arguments{
\item{x}{A numeric vector of x-coordinates.}

\item{y}{A numeric vector of y-coordinates.}

\item{cx}{A numeric values for the center x-coordinate.}

\item{cy}{A numeric values for the center y-coordinate.}

\item{angle_deg}{Rotation in degrees.}
}
\value{
Matrix of rotated x and y.
}
\description{
Rotate a point around another point in 2D.
}
\examples{
rotate_around(3, 4, 1, 2, 90)
}
